Call a notify function in Schema.__setattr__()
VObject will soon be sporting a new cache of older versions that will
need to be invalidated if a value changes.  To support this, it is
necessary for Schema to notify VObject when a value changes.  (We have
to have this in Schema because modifications could come from methods
declared on the Schema.)  This adds the necessary notification
mechanism.