Split SmartVersion class
This splits the SmartVersion class and its tests out into separate
files, as another step towards improving the layout.  It also adds
support to SmartVersion for Python 3 constructs, and completes the
implementation of the __getitem__() method by adding an as yet
unrealized call to the __vers_accessor__() method of the VObject.