Create SchemaProxy
This creates the SchemaProxy class, which becomes a base class for
VObject; many of the proxy-like methods from VObject are moved into
SchemaProxy.  This also creates ReadOnlyLazySchemaProxy, a special
proxy class which provides read-only access to a schema object; it
will be used to provide access to the data via an older version.