Make SmartVersion a true int
SmartVersion was a class that looks like an int, but the only
differences are the constructor and some methods not defined on
normal ints.  This means that SmartVersion can actually subclass
an int.  We have to switch __init__() out for __new__() and
eliminate the special integer methods, but otherwise it looks the
same, and now isinstance(version, int) will be True.