Cache upgrader Converters instance
Previously, any time we did __setstate__() from an older version, the
entire Converters instance had to be recomputed each time.  This could
be lengthy, so there is obviously the potential for precomputing or
caching.  Here we choose caching, on the assumption that an application
will only need to upgrade from one or two older versions of the object.
We also pre-cache the "identity" upgrader.